+++ src/extension.ts
@@ -2,7 +2,9 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode'
 import TaskTimer from './task-timer/task-timer'
-import { SidebarProvider } from './providers/sidebar-provider'
+import { SidebarMainProvider } from './providers/sidebar-main-provider'
+import { SidebarSummaryProvider } from './providers/sidebar-summary-provider'
+import SidebarState from './state/sidebar-state'
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -11,33 +13,43 @@ export function activate (context: vscode.ExtensionContext): void {
 
   const rootFilePath = context.globalStorageUri.fsPath
 
+  const sidebarState = new SidebarState(rootFilePath)
+
   // The command has been defined in the package.json file
   // Now provide the implementation of the command with registerCommand
   // The commandId parameter must match the command field in package.json
-  let disposable = vscode.commands.registerCommand('vscode-time-keeper.startTask', () => {
+  let disposable = vscode.commands.registerCommand('vscode-time-keeper.startTask', async () => {
     const timer = new TaskTimer(rootFilePath)
-    void timer.startTask()
+    await timer.startTask(sidebarState.calendarDate ?? new Date())
   })
 
   context.subscriptions.push(disposable)
 
-  disposable = vscode.commands.registerCommand('vscode-time-keeper.stopTask', () => {
+  disposable = vscode.commands.registerCommand('vscode-time-keeper.stopTask', async () => {
     const timer = new TaskTimer(rootFilePath)
-    void timer.stopTask()
+    await timer.stopTask(true, sidebarState.calendarDate ?? new Date())
   })
 
   context.subscriptions.push(disposable)
 
-  disposable = vscode.commands.registerCommand('vscode-time-keeper.generateWeeklyReport', () => {
+  disposable = vscode.commands.registerCommand('vscode-time-keeper.generateWeeklyReport', async () => {
     const timer = new TaskTimer(rootFilePath)
-    void timer.generateWeeklyReport()
+    if (sidebarState.calendarDate == null) {
+      await timer.generateWeeklyReport()
+      return
+    }
+    await timer.generateWeeklyReport(sidebarState.calendarDate)
   })
 
   context.subscriptions.push(disposable)
 
-  disposable = vscode.commands.registerCommand('vscode-time-keeper.editTimeLog', () => {
+  disposable = vscode.commands.registerCommand('vscode-time-keeper.editTimeLog', async () => {
     const timer = new TaskTimer(rootFilePath)
-    void timer.editTimeLog()
+    if (sidebarState.calendarDate == null) {
+      await timer.editTimeLog()
+      return
+    }
+    await timer.editTimeLog(sidebarState.calendarDate)
   })
 
   const settings = vscode.workspace.getConfiguration('TimeKeeper')
@@ -51,11 +63,20 @@ export function activate (context: vscode.ExtensionContext): void {
   }
 
   // Register the Sidebar Panel
-  const sidebarProvider = new SidebarProvider(context.extensionUri, rootFilePath)
+  const sidebarProvider = new SidebarMainProvider(context.extensionUri, sidebarState)
   context.subscriptions.push(
     vscode.window.registerWebviewViewProvider(
-      'time-keeper-sidebar',
+      'time-keeper-sidebar-main',
       sidebarProvider
+    )
+  )
+
+  // Register the Summary Panel
+  const summaryProvider = new SidebarSummaryProvider(context.extensionUri, sidebarState)
+  context.subscriptions.push(
+    vscode.window.registerWebviewViewProvider(
+      'time-keeper-summary',
+      summaryProvider
     )
   )
 }